fix change id type entities for booking

+++ hotel-backend/src/main/java/com/skypay/hotel/entity/Booking.java
@@ -15,7 +15,7 @@ import java.time.temporal.ChronoUnit;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Booking {
-    private Long bookingId;
+    private Integer bookingId;
     private Integer userId;
     private Integer roomNumber;
     private LocalDate checkIn;
